ac_fs: Re-work ac_fs_mkdir_p()
Use mkdirat(2) instead of mkdir(2), see openat(2) for why this is
desirable.

This does change the function prototype to add two arguments, a dirfd
which is a file descriptor for an opened directory, if the specified
path is relative, it will be created relative to the directory pointed
to by dirfd.

If the path is absolute then dirfd is ignored.

If path is relative and dirfd is specified as AC_FS_AT_FDCWD (a synonym
for AT_FDCWD) the path is created relative to the processes current
working directory, ala mkdir(2).

mode specifies the mode for the new directory.

Also fix a potential issue in that if we get an error that isn't EEXIST
from mkdirat then return -1 (error) rather than 0 (success).

mkdirat(2) was added to the Linux kernel in 2.6.16 (2006-03-20) and
glibc in 2.4 (2006-03-06).

Closes: #1
Signed-off-by: Andrew Clayton <andrew@digital-domain.net>