src/Makefile: Remove the glibc version check
We checked for the version of glibc to see if we needed to link with
-lrt for the clock_* functions. The need to link with -lrt was removed
in glibc 2.17 released in December 2012.

The main reason for this check was to work on CentOS 6, but that is EOL
and CentOS 7 comes with glibc 2.17 so just remove this check now, but
make it possible to still build on older glibc by setting LIBS=-lrt, e.g

    $ LIBS=-lrt make

will append -lrt to the LIBS variable. You can also do similar with
CFLAGS & LDFLAGS.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>