ac_fs: Fix ac_fs_copy()
Unfortunately, commit b8ebfc1 ("ac_fs/ac_fs_copy: Some minor code
refactor") caused some change in behaviour. We would only ever call the
first sync_file_range(2) and not the second or any of the
posix_fadvise(2) calls. It probably still worked, just not how it's
supposed to.

Due to the

  if (window == 0)
      continue;

check we never increment window and always short circuit round the loop.

Rather than continuing create a label at the end of the loop, increment
the window there and jump to it the first time round the loop.

We do this rather than revert the broken commit so as to keep the intent
of that change.

Fixes: b8ebfc1 ("ac_fs/ac_fs_copy: Some minor code refactor")
Signed-off-by: Andrew Clayton <andrew@digital-domain.net>