ac_net: Have ac_net_port_from_sa() return -1 on error
Currently 0 is used to indicate an error from ac_net_port_from_sa(),
however it is maybe possible[0] (though unlikely/unusual) that you could
have a port number of 0 in the various sockaddr structures.

It may be wise to cater for this situation.

Make ac_net_port_from_sa() return an int rather than a u16 and have it
return -1 rather than 0 on error.

[0]: https://daniel.haxx.se/blog/2014/10/25/pretending-port-zero-is-a-normal-one/

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>