ac_jsonw: Pass a const pointer into json_build_str()
In ac_jsonw_add_real() create a const char * that points to the format
character buffer and pass that const pointer into json_build_str()
rather than the format buffer directly.

Whether this makes anything more secure or not... but it can't hurt and
it's what json_build_str() wants.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>