ac_geo: Combine ellipsoid structure and ellipsoids[] array
We don't really need the actual ellipsoid structure definition so just
define it anonymously at the same time we create the ellipsoids[] array.

We also index initialise the array by ac_geo_ellipsoid_t this time as
it's important the order of the array matches the ac_geo_ellipsoid_t
enum values.

Signed-off-by: Andrew Clayton <ac@sigsegv.uk>