Add a __nonstring macro
Under GCC this is a wrapper around __attribute__((nonstring)) which is
used to mark character arrays as intentionally not being NUL terminated.

GCC 15 introduced a new warning, Wunterminated-string-initialization,
which will always warn on things like

  static char a[3] = "ABC";

as it lacks a terminating NUL byte.

However this is sometimes very much intentionally intended, where the
array is treated like a lookup table and only individual indices are
accessed.

The above attribute will quell this new warning (which is also enabled
with -Wextra).

Link: <https://gcc.gnu.org/git/?p=gcc.git;a=commit;h=44c9403ed1833ae71a59e84f9e37af3182be0df5>
Link: <https://gcc.gnu.org/git/gitweb.cgi?p=gcc.git;h=622968990beee7499e951590258363545b4a3b57>
Link: <https://gcc.gnu.org/bugzilla/show_bug.cgi?id=117178#c21>
Signed-off-by: Andrew Clayton <ac@sigsegv.uk>