include: Change __nonstring guard
Clang 21 added support for the -Wunterminated-string-initialization
warning as added in GCC 15.

Commit 4e5bb56 ("Add a __nonstring macro") added a macro around the
"nonstring" attribute when using GCC to quell this warning when desired.

Now that clang also has this warning we now get tripped up as we did in
GCC.

Change the guard around the macro to use __has_attribute() instead of
checking for the compiler as simply enabling also for clang will trigger
warnings from older clangs that don't have the "nonstring" attribute.

__has_attribute() is supported from GCC 5 and Clang 2.9, which should be
fine for us.

Signed-off-by: Andrew Clayton <ac@sigsegv.uk>